Use float and snprintf
It turns out that std::to_string() can't control the amount of decimal places. Use snprintf instead so we can control the number of decimal places. Use float to replace double because float uses less memory and we don't need the precision.

+++ components/webServer/include/Data2Web.h
@@ -14,13 +14,13 @@
 class historyObj
 {
 public:
-    double max;
-    double avg;
-    double min;
-    std::vector<double> valueHolder;
+    float max;
+    float avg;
+    float min;
+    std::vector<float> valueHolder;
 };
 
 char *getDataNow(long long time);
 char *getHostInfoJson();
 char *getDaysStats();
-void addToRecent(std::vector<double> liveData);
+void addToRecent(std::vector<float> liveData);
